Hide Amqp internals better
