--- amqp/src/main/scala/akka/stream/alpakka/amqp/package.scala
@@ -1,10 +1,0 @@
-/*
- * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
- */
-
-package akka.stream.alpakka
-
-package object amqp {
-  val Seq = scala.collection.immutable.Seq
-  type Seq[+A] = scala.collection.immutable.Seq[A]
-}
