+++ amqp/src/test/scala/akka/stream/alpakka/amqp/AmqpSpec.scala
@@ -5,6 +5,7 @@
 package akka.stream.alpakka.amqp
 
 import akka.actor.ActorSystem
+import akka.dispatch.ExecutionContexts
 import akka.stream.ActorMaterializer
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.{BeforeAndAfterAll, Matchers, WordSpec}
@@ -13,6 +14,7 @@ abstract class AmqpSpec extends WordSpec with Matchers with BeforeAndAfterAll wi
 
   implicit val system = ActorSystem(this.getClass.getSimpleName)
   implicit val materializer = ActorMaterializer()
+  implicit val executionContext = ExecutionContexts.sameThreadExecutionContext
 
   override protected def afterAll(): Unit =
     system.terminate()
